+++ src/ifcblenderexport/blenderbim/bim/module/cobie/__init__.py
@@ -1,0 +1,17 @@
+import bpy
+from . import ui, operator
+
+classes = (
+    operator.SelectCobieIfcFile,
+    operator.SelectCobieJsonFile,
+    operator.ExecuteIfcCobie,
+    ui.BIM_PT_cobie,
+)
+
+
+def register():
+    pass
+
+
+def unregister():
+    pass
